fix: dont die if semget returns 0
which is technically allowed, as -1 is the error value. The assumption here
is that when running inside a container, there are no other semaphores due
to isolation and thus the returned identifier is 0, whereas other deployments
would expect a non-null id simply because earlier users would already
have created a semaphore.

Signed-off-by: Alexander Bartolomey <occloxium@gmail.com>